receipt-bot/pyproject.toml
[project]
name = "receipt-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth>=2.49.1",
    "langchain-core>=1.2.21",
    "langchain-groq>=1.1.2",
    "pydantic>=2.12.5",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
    "structlog>=25.5.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
